Putting psychometrics code into the repository.
