+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/joe/examples/ExampleCompareFromFiles.java
@@ -19,11 +19,12 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.test.joes.examples;
+package edu.cmu.tetrad.algcomparison.joe.examples;
 
 import edu.cmu.tetrad.algcomparison.Comparison;
 import edu.cmu.tetrad.algcomparison.algorithm.Algorithms;
-import edu.cmu.tetrad.algcomparison.algorithm.external.ExternalAlgorithmPcalgPc;
+import edu.cmu.tetrad.algcomparison.algorithm.oracle.pattern.*;
+import edu.cmu.tetrad.algcomparison.score.FisherZScore;
 import edu.cmu.tetrad.algcomparison.statistic.*;
 import edu.cmu.tetrad.util.Parameters;
 
@@ -38,14 +39,17 @@ import edu.cmu.tetrad.util.Parameters;
  *
  * @author jdramsey
  */
-public class ExampleCompareFromFilesSimple {
+public class ExampleCompareFromFiles {
     public static void main(String... args) {
         Parameters parameters = new Parameters();
 
         // Can leave the simulation parameters out since
         // we're loading from file here.
-        parameters.set("alpha", 1e-3);
+        parameters.set("alpha", 1e-8);
         parameters.set("numRuns", 10);
+        parameters.set("penaltyDiscount", 4);
+        parameters.set("useMaxPOrientationHeuristic", true);
+        parameters.set("maxPOrientationMaxPathLength", 3);
 
         Statistics statistics = new Statistics();
 
@@ -55,12 +59,13 @@ public class ExampleCompareFromFilesSimple {
         statistics.add(new AdjacencyRecall());
         statistics.add(new ArrowheadPrecision());
         statistics.add(new ArrowheadRecall());
-        statistics.add(new MathewsCorrAdj());
-        statistics.add(new MathewsCorrArrow());
-        statistics.add(new F1Adj());
-        statistics.add(new F1Arrow());
-        statistics.add(new SHD());
-        statistics.add(new ElapsedTime());
+        statistics.add(new PercentBidirectedEdges());
+//        statistics.add(new MathewsCorrAdj());
+//        statistics.add(new MathewsCorrArrow());
+//        statistics.add(new F1Adj());
+//        statistics.add(new F1Arrow());
+//        statistics.add(new SHD());
+//        statistics.add(new ElapsedTime());
 
         statistics.setWeight("AP", 1.0);
         statistics.setWeight("AR", 0.5);
@@ -70,16 +75,39 @@ public class ExampleCompareFromFilesSimple {
         Algorithms algorithms = new Algorithms();
 
 //        algorithms.add(new Pc(new FisherZ()));
-        algorithms.add(new ExternalAlgorithmPcalgPc("rgraph"));
+//        algorithms.add(new PcStable(new FisherZ()));
+//        algorithms.add(new PcStableMax(new FisherZ(), false));
+//        algorithms.add(new Cpc(new FisherZ()));
+//        algorithms.add(new CpcStable(new FisherZ()));
+        algorithms.add(new Fges(new FisherZScore()));
+//
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.original", "PC from pcalg, original adjacency search, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.stable", "PC from pcalg, stable adjacency search, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.stable.fast", "PC from pcalg, stable fast adjacency search, alpha = 0.001"));
+//
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.relaxed", "PC from pcalg, relaxed, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.rand", "PC from pcalg, rand, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.retry", "PC from pcalg, retry, alpha = 0.001"));
+//
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.NAdelete.FAlSE", "PC from pcalg, NAdelete=FALSE, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.conservative.FAlSE", "PC from pcalg, conservative=FALSE, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.conservative.TRUE", "PC from pcalg, conservative=TRUE, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.maj.rule.TRUE", "PC from pcalg, maj.rule=TRUE, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.solve.confl.FALSE", "PC from pcalg, solve.confl=FALSE, alpha = 0.001"));
+//        algorithms.add(new ExternalAlgorithmPcalgPc("pc.solve.confl.TRUE", "PC from pcalg, solve.confl=TRUE, alpha = 0.001"));
+
+//        algorithms.add(new ExternalAlgorithmBnlearnMmhc("bnlearn.mmhc", "MMHC from bnlearn, alpha = 0.001"));
 
         Comparison comparison = new Comparison();
         comparison.setShowAlgorithmIndices(true);
         comparison.setShowSimulationIndices(true);
-        comparison.setSortByUtility(true);
-        comparison.setShowUtilities(true);
-        comparison.setParallelized(true);
+        comparison.setSortByUtility(false);
+        comparison.setShowUtilities(false);
+        comparison.setSaveGraphs(true);
 
-        comparison.compareFromFiles("comparison_simple", algorithms, statistics, parameters);
+        comparison.compareFromFiles("/Users/user/comparison-data/condition_1",
+                "/Users/user/causal-comparisons/condition_1",
+                algorithms, statistics, parameters);
     }
 }
