+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/joe/examples/ExampleSave2.java
@@ -19,7 +19,7 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.test.joes.examples;
+package edu.cmu.tetrad.algcomparison.joe.examples;
 
 import edu.cmu.tetrad.algcomparison.Comparison;
 import edu.cmu.tetrad.algcomparison.graph.RandomForward;
